refactor: rename AddFriendViewController to FriendListViewController
